Vertx integration

+++ src/main/java/org/sai/predmod/entity/PredictiveModel.java
@@ -33,7 +33,7 @@ public class PredictiveModel {
     private long lastTrainedDateTime;
 
     @Column
-    private long lastTrainingTimeTookInSeconds;
+    private double lastTrainingTimeTookInSeconds;
 
     @Column
     private long trainingDatasetSize;
